Consolidate hooks
